Minor fixes for Linux build

+++ sources/wxsha2.cpp
@@ -2,7 +2,7 @@
 #include <fstream>
 #include "wxsha2.h"
 
-const unsigned int wxSHA256::sha256_k[64] = //UL = uint32
+const uint32_t wxSHA256::sha256_k[64] = //UL = uint32
             {0x428a2f98, 0x71374491, 0xb5c0fbcf, 0xe9b5dba5,
              0x3956c25b, 0x59f111f1, 0x923f82a4, 0xab1c5ed5,
              0xd807aa98, 0x12835b01, 0x243185be, 0x550c7dc3,
@@ -19,7 +19,7 @@ const unsigned int wxSHA256::sha256_k[64] = //UL = uint32
              0x391c0cb3, 0x4ed8aa4a, 0x5b9cca4f, 0x682e6ff3,
              0x748f82ee, 0x78a5636f, 0x84c87814, 0x8cc70208,
              0x90befffa, 0xa4506ceb, 0xbef9a3f7, 0xc67178f2};
-const unsigned long long wxSHA512::sha512_k[80] = //ULL = uint64
+const uint64_t wxSHA512::sha512_k[80] = //ULL = uint64
             {0x428a2f98d728ae22ULL, 0x7137449123ef65cdULL,
              0xb5c0fbcfec4d3b2fULL, 0xe9b5dba58189dbbcULL,
              0x3956c25bf348b538ULL, 0x59f111f1b605d019ULL,
